backend/requirements.txt
Flask==2.2.2
PyJWT==2.7.0
pymongo==4.3.3
pypokedex==1.6.0
python-dotenv==1.0.0
bcrypt==3.1.7
waitress==2.1.2
Flask-Cors==3.0.10